+++ pygls/types.py
@@ -686,6 +686,15 @@ class Location:
         self.uri = uri
         self.range = range
 
+    def __eq__(self, other):
+        return (
+            isinstance(other, Location)
+            and self.uri == other.uri
+            and self.range == other.range)
+
+    def __repr__(self):
+        return "{}:{}".format(self.uri, self.range)
+
 
 class LocationLink:
     def __init__(self,
@@ -753,9 +762,10 @@ class Position:
         self.character = character
 
     def __eq__(self, other):
-        if self.line == other.line and self.character == other.character:
-            return True
-        return False
+        return (
+            isinstance(other, Position)
+            and self.line == other.line
+            and self.character == other.character)
 
     def __ge__(self, other):
         line_gt = self.line > other.line
@@ -823,6 +833,15 @@ class Range:
     def __init__(self, start: Position, end: Position):
         self.start = start
         self.end = end
+
+    def __eq__(self, other):
+        return (
+            isinstance(other, Range)
+            and self.start == other.start
+            and self.end == other.end)
+
+    def __repr__(self):
+        return '{}-{}'.format(self.start, self.end)
 
 
 class ReferenceContext: